% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{mpod_spectrum}
\alias{mpod_spectrum}
\title{Optical Density of the Macular Pigment}
\format{
A data frame with 196 rows and 2 columns:
\describe{
\item{Wavelength}{Numeric. Wavelength in nanometers (nm).}
\item{macss_1.optical_density}{Numeric. Measured optical density of the macular pigment.}
}
}
\source{
Bone, R. A., Landrum, J. T., & Cains, A. (1992). Optical density spectra of the macular pigment in vivo and in vitro. Vision Research, 32, 105-110.

Stockman, A., Sharpe, L. T., & Fach, C. C. (1999). The spectral sensitivity of the human short-wavelength cones. Vision Research, 39, 2901-2927.

www.cvrl.org
}
\usage{
data(mpod_spectrum)
}
\description{
This dataset contains the spectral optical density of the macular pigment as a function of wavelength.
Measurements were taken from 390 nm to 780 nm in 2-nm increments.
}
\details{
The macular pigment is a yellowish pigment located in the macula, the central area of the retina.
It primarily absorbs short-wavelength (blue) light, thereby protecting the retina from phototoxic damage.
The optical density indicates how strongly the pigment absorbs light at a given wavelength.
A value of 0 means no absorption, while higher values indicate stronger absorption.
}
\examples{
data(mpod_spectrum)
head(mpod_spectrum)
plot(mpod_spectrum$Wavelength,
     mpod_spectrum$macss_1.optical_density,
     type = "l", xlab = "Wavelength (nm)", ylab = "Optical Density",
     main = "Optical Density of the Macular Pigment")

}
\keyword{datasets}
