% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{photoreceptor_spectral_sensitivities}
\alias{photoreceptor_spectral_sensitivities}
\title{Spectral Sensitivity Functions of Human Photoreceptors}
\format{
A data frame with 196 rows and 5 columns:
\describe{
\item{melanopsin}{Numeric. Relative spectral sensitivity of melanopsin-containing ipRGCs.}
\item{rod}{Numeric. Relative spectral sensitivity of rods.}
\item{scone}{Numeric. Relative spectral sensitivity of short-wavelength-sensitive cones (S-cones).}
\item{mcone}{Numeric. Relative spectral sensitivity of medium-wavelength-sensitive cones (M-cones).}
\item{lcone}{Numeric. Relative spectral sensitivity of long-wavelength-sensitive cones (L-cones).}
}
}
\source{
S-cones: Stockman, A., Sharpe, L. T., & Fach, C. C. (1999). The spectral sensitivity of the human short-wavelength cones. Vision Research, 39, 2901-2927.

L-/M-cones: Stockman, A., & Sharpe, L. T. (2000). Spectral sensitivities of the middle- and long-wavelength sensitive cones derived from measurements in observers of known genotype. Vision Research, 40, 1711-1737.

Rods: Wyszecki, G., & Stiles, W. S. (1982). Color Science: concepts and methods, quantitative data and formulae. (2nd ed.). New York: Wiley.

Melanopsin: DOI: 10.25039/CIE.DS.vqqhzp5a

www.cvrl.org
}
\usage{
data(photoreceptor_sensitivities)
}
\description{
This dataset contains the normalized spectral sensitivity functions for the five main human photoreceptor types:
melanopsin-containing intrinsically photosensitive retinal ganglion cells (ipRGCs), rods, S-cones, M-cones, and L-cones.
The data represent the relative sensitivity of each photoreceptor type across the visible spectrum.
}
\details{
The spectral sensitivities are normalized and typically derived from physiological measurements or established template functions.
These functions are widely used in vision science, chronobiology, and lighting research to model the response of the human visual and non-visual system to light stimuli.
}
\examples{
data(photoreceptor_sensitivities)
head(photoreceptor_sensitivities)
matplot(photoreceptor_sensitivities, type = "l", lty = 1,
        col = c("black", "grey", "blue", "green", "red"),
        xlab = "Index (arbitrary, e.g., wavelength)", ylab = "Relative Sensitivity",
        main = "Spectral Sensitivity Functions of Human Photoreceptors")

}
\keyword{datasets}
