% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{vlambda}
\alias{vlambda}
\title{CIE 1924 Photopic Luminous Efficiency Function (V-lambda)}
\format{
A data frame with 196 rows and 3 columns:
\describe{
\item{wavelength_nm}{Numeric. Wavelength in nanometers (nm).}
\item{frequency_THz}{Numeric. Frequency in terahertz (THz) corresponding to each wavelength.}
\item{vlambda}{Numeric. Value of the photopic luminous efficiency function V(λ) at each wavelength.}
}
}
\source{
www.cvrl.org
}
\usage{
data(vlambda_cie1924)
}
\description{
This dataset provides the CIE 1924 photopic luminous efficiency function V(λ), together with the corresponding wavelengths and frequencies.
The data cover the visible spectrum from 390 nm to 780 nm in 2-nm increments.
}
\details{
The photopic luminous efficiency function V(λ) describes the average spectral sensitivity of human visual perception of brightness under well-lit (photopic) conditions.
It is a key standard function in color science and photometry, and is used to convert spectral power distributions to luminous quantities such as lumens.
}
\examples{
data(vlambda_cie1924)
head(vlambda_cie1924)
plot(vlambda_cie1924$wavelength_nm, vlambda_cie1924$vlambda, type = "l",
     xlab = "Wavelength (nm)", ylab = "V(lambda)",
     main = "CIE 1924 Photopic Luminous Efficiency Function")

}
\keyword{datasets}
